ODIN-22 remove canvas and library service, move canvas to viewport component, remove cicle DI
