+++ src/app/services/index.service.ts
@@ -1,7 +1,6 @@
 export { AppService } from './app.service';
-export { CanvasService } from './canvas.service';
 export { EngineService } from './engine.service';
-export { LibraryService } from './library.service';
 export { LogService } from './log.service';
 export { SessionService } from './session.service';
 export { WindowService as WindowRefService } from './window.service';
+
