+++ src/app/ui/builder-panel/builder-panel.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { select, Store } from '@ngrx/store';
 import { Mesh, VertexData } from 'babylonjs';
 import { Container } from 'src/app/engine/common/Container';
+import { EngineService } from 'src/app/services/index.service';
 import { AppState } from '../../store/app.reducer';
-import { EngineService } from './../../services/index.service';
+import { ViewportComponent } from './../viewport/viewport.component';
 
 @Component({
   selector: 'builder-panel',
@@ -14,14 +15,14 @@ export class BuilderPanelComponent implements OnInit {
 
   public current: Container;
   constructor(
-    public es: EngineService,
+    private engineServ: EngineService,
     public store: Store<AppState>
   ) { }
 
   ngOnInit(): void {
     this.store.pipe(select('engine')).subscribe(en => {
       if (en.UUIDCsSelected.length == 0) { this.current = undefined; return; }
-      let container = this.es.getContainerFromUUID(en.UUIDCsSelected[0]);
+      let container = this.engineServ.getContainerFromUUID(en.UUIDCsSelected[0]);
       if (!(container.type instanceof Mesh)) this.current = undefined; else this.current = container;
     });
   }
@@ -29,6 +30,6 @@ export class BuilderPanelComponent implements OnInit {
   updateMesh() {
     let vd: VertexData = this.current.rebuildMesh(this.current.panel.values);
     vd.applyToMesh(<Mesh>this.current.type, true);
-    this.es.getCanvasHelper().setBoundingBoxText(<Mesh>(this.current.type));
+    ViewportComponent.setBoundingBoxText(<Mesh>(this.current.type));
   }
 }
