+++ src/app/ui/viewport/viewport.component.scss
@@ -1,0 +1,18 @@
+.engine-wrapper {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 300px;
+  z-index: 0;
+
+  #renderCanvas {
+    width: 100%;
+    height: 100%;
+    touch-action: none;
+  }
+
+  #renderCanvas:focus {
+    outline: none;
+  }
+}
