+++ src/styles.scss
@@ -103,33 +103,6 @@ body {
   text-align: center;
 }
 
-.engine-wrapper {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 300px;
-  z-index: 0;
-
-  #renderCanvas {
-    width: 100%;
-    height: 100%;
-    touch-action: none;
-  }
-
-  #renderCanvas:focus {
-    outline: none;
-  }
-}
-
-.ui-wrapper {
-  position: absolute;
-
-  * {
-    z-index: 1;
-  }
-}
-
 /* general */
 .float-left {
   float: left;
